+++ packages/components/src/components/AppBar.stories.js
@@ -6,7 +6,7 @@ import { AppBar, Text } from '@wingscms/components';
 
 const props = ({ position = AppBar.Position.TOP } = {}) => ({
   position: select('position', AppBar.Position, position),
-  open: boolean('open', true),
+  hide: boolean('hide', false),
 });
 
 const ExampleContent = styled(Text)`
@@ -17,12 +17,9 @@ const ExampleContent = styled(Text)`
 
 export default () => <AppBar {...props()} />;
 
-export const withContent = () => {
-  faker.seed(1);
-  return (
-    <>
-      <AppBar {...props()} />
-      <ExampleContent>{faker.lorem.paragraphs(50)}</ExampleContent>
-    </>
-  );
-};
+export const withContent = () => (
+  <>
+    <AppBar {...props()} />
+    <ExampleContent>{faker.lorem.paragraphs(50)}</ExampleContent>
+  </>
+);