+++ packages/react/src/components/cards/CTACard.stories.js
@@ -1,16 +1,12 @@
 import { CTACard } from '@wingscms/react';
-import { CallToAction } from '@wingscms/components';
-import { boolean, select, text } from '@storybook/addon-knobs/react';
 import { contentWrap, image } from '../../../../../utils';
 
 const props = ({ backgroundImage = null } = {}) => ({
-  align: select('align', CallToAction.Align, CallToAction.Align.LEFT),
-  backgroundImage: text('backgroundImage', backgroundImage),
-  buttonText: text('buttonText', 'Sign up now'),
-  buttonUrl: text('buttonUrl', 'http://example.com'),
-  reveal: boolean('reveal', true),
-  text: text('text', 'This is an introduction to the first section of the article'),
-  title: text('title', 'Section One'),
+  backgroundImage,
+  buttonText: 'Sign up now',
+  buttonUrl: 'http://example.com',
+  text: 'This is an introduction to the first section of the article',
+  title: 'Section One',
 });
 
 export default () => CTACard.render(props());