+++ packages/mobiledoc-renderer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://github.com/wingscms/wings/compare/@wingscms/mobiledoc-renderer@0.1.1...@wingscms/mobiledoc-renderer@1.0.0) (2020-05-26)
+
+
+### Features
+
+* add Link and styles to component library ([#437](https://github.com/wingscms/wings/issues/437)) ([d64088f](https://github.com/wingscms/wings/commit/d64088f45a603abff5533c5c6f606676dcf28e61))
+
+
+
+
+
 ## 0.1.1 (2020-05-12)